file extension was changed from js to ts
