+++ server/index.ts
@@ -1,0 +1,20 @@
+const express = require("express");
+const cors = require("cors");
+const connectDB = require("./config/db");
+const routes = require("./routes/todo");
+
+const app = express();
+connectDB();
+
+
+app.use(cors({ origin: true, credentials: true }));
+app.use(express.json({ extended: false }));
+
+app.use("/api/todo", routes);
+app.get("/", (req, res) => res.send("Server up and running"));
+
+const PORT = process.env.PORT || 8000;
+
+app.listen(PORT, () => {
+  console.log(`server is running on port ${PORT}`);
+});
