+++ server/models/todo.ts
@@ -1,0 +1,17 @@
+// models/todo.js
+const mongoose = require("mongoose");
+
+const TodoSchema = new mongoose.Schema({
+  text: {
+    type: String,
+    required: true,
+  },
+  complete: {
+    type: Boolean,
+    required: true,
+  },
+});
+
+const Todo = mongoose.model("todo", TodoSchema);
+
+module.exports = Todo;
